requirements.txt
# Bot framework
python-telegram-bot[job-queue]==22.6

# HTTP clients
aiohttp>=3.9
httpx>=0.27

# HTML parsing (para scraping do catalogo)
beautifulsoup4>=4.12
lxml>=5.0

# Outras dependências úteis
python-dotenv>=1.0
pycryptodomex>=3.20
